requirements.txt
neo4j==5.24.0
openai
google-generativeai==0.8.3
pydantic==2.9.2
termcolor==2.4.0
python-dotenv==1.0.1
tiktoken==0.8.0
pytest==8.3.3
pytest-asyncio==0.24.0
asyncio==3.4.3
langchain-openai
langsmith
numpy>=1.24.0
tabulate>=0.9.0
